+++ src/ace_time/BasicZoneProcessor.h
@@ -210,8 +210,7 @@ class BasicZoneProcessorTemplate: public ZoneProcessor {
 
     uint32_t getZoneId(bool followLink = false) const override {
       ZIB zib = (isLink() && followLink)
-          ? mBrokerFactory->createZoneInfoBroker(
-                (uintptr_t) mZoneInfoBroker.targetZoneInfo())
+          ? mZoneInfoBroker.targetZoneInfo()
           : mZoneInfoBroker;
       return zib.zoneId();
     }
@@ -315,8 +314,7 @@ class BasicZoneProcessorTemplate: public ZoneProcessor {
 
     void printNameTo(Print& printer, bool followLink = false) const override {
       ZIB zib = (isLink() && followLink)
-          ? mBrokerFactory->createZoneInfoBroker(
-                (uintptr_t) mZoneInfoBroker.targetZoneInfo())
+          ? mZoneInfoBroker.targetZoneInfo()
           : mZoneInfoBroker;
       zib.printNameTo(printer);
     }
@@ -324,8 +322,7 @@ class BasicZoneProcessorTemplate: public ZoneProcessor {
     void printShortNameTo(Print& printer, bool followLink = false)
         const override {
       ZIB zib = (isLink() && followLink)
-          ? mBrokerFactory->createZoneInfoBroker(
-                (uintptr_t) mZoneInfoBroker.targetZoneInfo())
+          ? mZoneInfoBroker.targetZoneInfo()
           : mZoneInfoBroker;
       zib.printShortNameTo(printer);
     }